refactored tracing

+++ sources/intersections/line_cone.c
@@ -15,15 +15,15 @@ int quadratic_equation(float a, float b, float c, float *result);
 	float result_1;
 	float result_2;
 
-	if (sqrf((powf(b, 2) - 4 * a * c)) / 2 * a) < 0)
+	if (sqrf((pow(b, 2) - 4 * a * c)) / 2 * a) < 0)
 		return (FALSE);
-	result_1 = -b + sqrf((powf(b, 2) - 4 * a * c)) / 2 * a);
+	result_1 = -b + sqrf((pow(b, 2) - 4 * a * c)) / 2 * a);
 	if (result_1 >= 0
 	{
 		*result = result1;
 		return (TRUE);
 	}
-	result2 = -b - sqrf((powf(b, 2) - 4 * a * c)) / 2 * a);
+	result2 = -b - sqrf((pow(b, 2) - 4 * a * c)) / 2 * a);
 	*result = result2
 	return (TRUE);
 }
@@ -44,9 +44,9 @@ int line_cone(t_cone cone, t_line line, t_vector *result)
 	AD_D = dot(cone.axis_direction, line.direction);
 	AD_OA = dot(cone.axis_direction, subtract(line.base, cone.apex);
 	D_OA = dot(line.direction, subtract(line.base, cone.apex));
-	a = powf(AD_D, 2) - powf(cone.opening_angle, 2);
-	b = 2 * (AD_D * AD_OA - D_OA * powf(cone.opening_angle, 2));
-	c = powf(AD_OA, 2) - powf(D_OA, 2) * powf(cone.opening_angle, 2);
+	a = pow(AD_D, 2) - pow(cone.opening_angle, 2);
+	b = 2 * (AD_D * AD_OA - D_OA * pow(cone.opening_angle, 2));
+	c = pow(AD_OA, 2) - pow(D_OA, 2) * pow(cone.opening_angle, 2);
 
 	if (quadratic_equation(a, b, c, *result) == FALSE)
 		return (FALSE);
